pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Timur Galiev <timurgravirovsky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
pytils = "^0.4.1"
pillow = "^9.3.0"
drf-spectacular = "^0.25.1"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
